add check for file is null in test cleanup and up timeout to firefox browser

+++ TestDeveloper/TestProject/Tests/PortalFunctionsTest.cs
@@ -41,7 +41,10 @@ namespace TestProject.Tests
         [Priority(1)]
         public void DeleteFile()
         {
-            FileUtil.DeleteFileWithWait(_pathTestAttachment);
+            if (_pathTestAttachment != null)
+            {
+                FileUtil.DeleteFileWithWait(_pathTestAttachment);
+            }
         }
 
         [TestMethod]
